// src/StaticData/profiles/profile.interface.ts
export interface UserProfile {
  id: number;
  userId: number;
  username: string;
  email: string;
  firstname: string;
  surname: string;
  role: number;
  group: string;
  faculty: number;
  questions: number;
  answers: number;
  avatar_url: string;
  status: string;
  socialLinks: string;
  created_at: string;
  updated_at: string;
}